Fix server image build by moving the client config schema
#301 made the admin server import internal/client/config to validate team
templates against the client's own rules. .dockerignore excludes
internal/client from the server image, so the Docker build fails on main even
though `go build ./...` on a full checkout passes.

The config schema is shared: the server generates and now validates client
configs. Move the package to internal/clientconfig, which is outside the
excluded tree. Callers alias it as `config`, so only the import line changes.

CI only built the web-builder stage, so it never compiled the server the way
a release does. Build the whole image instead, which is the check that would
have caught this.
