+++ AGENTS.md
@@ -3,6 +3,8 @@
 ## Project Structure & Module Organization
 The repo root is flattened. Go entrypoints live in `cmd/portr` and `cmd/portrd`, with shared application code under `internal/`. Use `internal/server/admin` for admin server work, `internal/server/admin/web-v2` for the React + shadcn admin UI, and `internal/client/dashboard/ui-v2` for the client dashboard. Keep integration coverage in `tests/`, database changes in `migrations/`, and product docs in `docs-v2/`. Do not add new code under the legacy `tunnel/` path.
 
+`internal/client` and `cmd/portr` are excluded from the server image by `.dockerignore`, so nothing under `internal/server` or `cmd/portrd` may import them; the build only fails inside Docker. The client config schema is shared by both sides and lives in `internal/clientconfig` for that reason.
+
 ## Build, Test, and Development Commands
 Use `make buildgo` to compile repo-owned Go packages and `make testgo` before opening a PR. These targets exclude frontend `node_modules`, which can contain Go files after dependency installation. Build the CLI with `make buildcli`, which outputs `./portr`.
