+++ cmd/portr/start.go
@@ -9,9 +9,9 @@ import (
 	"time"
 
 	"github.com/amalshaji/portr/internal/client/client"
-	"github.com/amalshaji/portr/internal/client/config"
 	"github.com/amalshaji/portr/internal/client/dashboard"
 	"github.com/amalshaji/portr/internal/client/db"
+	config "github.com/amalshaji/portr/internal/clientconfig"
 	"github.com/urfave/cli/v2"
 )
