+++ internal/client/client/stub_tunnel_test.go
@@ -6,7 +6,7 @@ import (
 	"net/http"
 	"testing"
 
-	clientcfg "github.com/amalshaji/portr/internal/client/config"
+	clientcfg "github.com/amalshaji/portr/internal/clientconfig"
 	"github.com/amalshaji/portr/internal/constants"
 )
