+++ internal/client/ssh/hostkey.go
@@ -10,7 +10,7 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/amalshaji/portr/internal/client/config"
+	config "github.com/amalshaji/portr/internal/clientconfig"
 	"github.com/charmbracelet/log"
 	"golang.org/x/crypto/ssh"
 )
