+++ internal/client/ssh/logging_test.go
@@ -8,8 +8,8 @@ import (
 	"strings"
 	"testing"
 
-	clientcfg "github.com/amalshaji/portr/internal/client/config"
 	clientdb "github.com/amalshaji/portr/internal/client/db"
+	clientcfg "github.com/amalshaji/portr/internal/clientconfig"
 	"github.com/glebarez/sqlite"
 	"gorm.io/gorm"
 	"gorm.io/gorm/logger"
