+++ internal/clientconfig/config.go
@@ -1,4 +1,4 @@
-package config
+package clientconfig
 
 import (
 	"errors"
