+++ internal/clientconfig/configfile.go
@@ -1,4 +1,4 @@
-package config
+package clientconfig
 
 import (
 	"bytes"
