+++ internal/clientconfig/template_test.go
@@ -1,4 +1,4 @@
-package config
+package clientconfig
 
 import (
 	"strings"
