+++ tests/tunnel_data_flow_test.go
@@ -16,9 +16,9 @@ import (
 	"testing"
 	"time"
 
-	clientconfig "github.com/amalshaji/portr/internal/client/config"
 	clientdb "github.com/amalshaji/portr/internal/client/db"
 	clientssh "github.com/amalshaji/portr/internal/client/ssh"
+	clientconfig "github.com/amalshaji/portr/internal/clientconfig"
 	"github.com/amalshaji/portr/internal/constants"
 	serverconfig "github.com/amalshaji/portr/internal/server/config"
 	serverdb "github.com/amalshaji/portr/internal/server/db"
